Revert the "Revert "19.10.1 - Dev to staging""

+++ src/app/service/pnr.service.ts
@@ -112,9 +112,11 @@ export class PnrService {
             if (type === 'RI') {
                 remarksList = this.pnrObj.riElements;
             }
-            for (const rm of remarksList) {
-                if (rm.freeFlowText.indexOf(searchText) === 0) {
-                    return rm.elementNumber;
+            if (remarksList) {
+                for (const rm of remarksList) {
+                    if (rm.freeFlowText.indexOf(searchText) === 0) {
+                        return rm.elementNumber;
+                    }
                 }
             }
         }
@@ -1040,7 +1042,7 @@ export class PnrService {
         return matrixReceipts;
     }
 
-    getSegmentModel(freetext, index, type) {
+    getSegmentModel(freetext, index, type, lineNo) {
         let segmentModel: PassiveSegmentsModel;
         segmentModel = new PassiveSegmentsModel();
 
@@ -1067,6 +1069,7 @@ export class PnrService {
             if (match !== null) {
                 segmentModel.isNew = false;
                 segmentModel.segmentNo = index;
+                segmentModel.lineNo = lineNo;
                 segmentModel.segmentType = match.groups.type;
                 if (!match.groups.type) {
                     segmentModel.segmentType = type;
@@ -1080,6 +1083,7 @@ export class PnrService {
                 segmentModel.arrivalDate = match.groups.arrdate;
                 segmentModel.arrivalTime = match.groups.arrtime;
                 segmentModel.confirmationNo = match.groups.conf;
+
             } else if (type === 'HTL') {
                 segmentModel.segmentType = type;
             }
@@ -1087,11 +1091,12 @@ export class PnrService {
         }
     }
 
-    getAirSegmentModel(element, index) {
+    getAirSegmentModel(element, index, lineNo) {
         let segmentModel: PassiveSegmentsModel;
         segmentModel = new PassiveSegmentsModel();
         segmentModel.isNew = false;
         segmentModel.segmentNo = index;
+        segmentModel.lineNo = lineNo;
         segmentModel.segmentType = element.segmentType;
         segmentModel.flightNumber = element.flightNumber;
         segmentModel.classService = element.classservice;
@@ -1105,11 +1110,12 @@ export class PnrService {
         return segmentModel;
     }
 
-    getHotelSegmentModel(element, index, freetext) {
+    getHotelSegmentModel(element, index, freetext, lineNo) {
         let segmentModel: PassiveSegmentsModel;
         segmentModel = new PassiveSegmentsModel();
         segmentModel.isNew = false;
         segmentModel.segmentNo = index;
+        segmentModel.lineNo = lineNo;
         segmentModel.segmentType = element.segmentType;
         segmentModel.departureDate = this.formatDate(element.deptdate);
         segmentModel.departureCity = element.cityCode;
@@ -1136,13 +1142,13 @@ export class PnrService {
             switch (element.segmentType) {
                 case 'MIS':
                 case 'CAR':
-                    pSegment.push(this.getSegmentModel(element.freetext, index, element.segmentType));
+                    pSegment.push(this.getSegmentModel(element.freetext, index, element.segmentType, element.lineNo));
                     break;
                 case 'AIR':
-                    pSegment.push(this.getAirSegmentModel(element, index));
+                    pSegment.push(this.getAirSegmentModel(element, index, element.lineNo));
                     break;
                 case 'HTL':
-                    pSegment.push(this.getHotelSegmentModel(element, index, element.freetext));
+                    pSegment.push(this.getHotelSegmentModel(element, index, element.freetext, element.lineNo));
                     break;
             }
         });
@@ -1496,7 +1502,7 @@ export class PnrService {
 
 
     public getTkLineDescription(): string {
-        if (this.pnrObj.tkElements && this.pnrObj.tkElements[0]) {
+        if (this.pnrObj.tkElements && this.pnrObj.tkElements[0] && this.pnrObj.tkElements[0].freeFlowText) {
             return this.pnrObj.tkElements[0].freeFlowText.trim();
         }
         return '';
